requirements.txt
transformers
torch
fastapi
uvicorn
requests
psutil
python-dotenv
huggingface_hub
llama-stack
Pillow
